refactor: improve context menu

+++ frontend/src/components/common/SavedQueries/SavedQueryItem/SavedQueryContextMenu/SavedQueryContextMenu.tsx
@@ -1,12 +1,12 @@
 import api from '@/api';
-import CustomIcon from '@/components/base/CustomIcon/CustomIcon';
+import ContextMenu from '@/components/base/ContextMenu/ContextMenu';
+import { MenuType } from '@/components/base/ContextMenu/types';
 import { TabMode } from '@/core/enums';
 import { useCopyToClipboard } from '@/hooks';
 import useAPI from '@/hooks/useApi.hook';
 import locales from '@/locales';
 import { useConfirmModalStore } from '@/store/confirmModal/confirmModal.store';
 import { useTabStore } from '@/store/tabStore/tab.store';
-import { Box, Menu, MenuItem, Stack } from '@mui/material';
 import { toast } from 'sonner';
 import { SavedQueryContextMenuProps } from '../../types';
 
@@ -27,7 +27,9 @@ export default function SavedQueryContextMenu({
   });
 
   const handleDelete = async () => {
-    onClose();
+    if (pendingDelete) {
+      return;
+    }
     showModal(locales.delete_action, locales.query_saved_delete_confirm, async () => {
       try {
         await deleteSavedQuery(query.id);
@@ -45,50 +47,39 @@ export default function SavedQueryContextMenu({
       toast.success(locales.database_delete_success);
     } catch (error) {
       console.log('🚀 ~ handleCopy ~ error:', error);
-    } finally {
-      onClose();
     }
   };
 
   const handleRun = () => {
     const name = query.name.slice(0, 10);
     addTab(name, TabMode.Query, query.query);
-    onClose();
   };
 
-  return (
-    <Box>
-      <Menu
-        open={contextMenu !== null}
-        onClose={onClose}
-        anchorReference='anchorPosition'
-        anchorPosition={contextMenu !== null ? { top: contextMenu.mouseY, left: contextMenu.mouseX } : undefined}
-      >
-        <MenuItem onClick={handleRun}>
-          <Stack width={'100%'} alignItems={'center'} justifyContent={'space-between'} direction={'row'}>
-            {locales.run}
-            <CustomIcon type='play' />
-          </Stack>
-        </MenuItem>
-        <MenuItem onClick={handleCopy}>
-          <Stack width={'100%'} alignItems={'center'} justifyContent={'space-between'} direction={'row'}>
-            {locales.copy}
-            <CustomIcon type='copy' />
-          </Stack>
-        </MenuItem>
-        <MenuItem onClick={onChange}>
-          <Stack width={'100%'} alignItems={'center'} justifyContent={'space-between'} direction={'row'}>
-            {locales.rename}
-            <CustomIcon type='pen' />
-          </Stack>
-        </MenuItem>
-        <MenuItem onClick={handleDelete}>
-          <Stack width={'100%'} alignItems={'center'} justifyContent={'space-between'} direction={'row'}>
-            {locales.delete}
-            <CustomIcon type='delete' />
-          </Stack>
-        </MenuItem>
-      </Menu>
-    </Box>
-  );
+  const menu: MenuType[] = [
+    {
+      name: locales.run,
+      icon: 'play',
+      action: handleRun,
+      closeAfterAction: true
+    },
+    {
+      name: locales.copy,
+      icon: 'copy',
+      action: handleCopy,
+      closeAfterAction: true
+    },
+    {
+      name: locales.rename,
+      icon: 'pen',
+      action: onChange
+    },
+    {
+      name: locales.delete,
+      icon: 'delete',
+      action: handleDelete,
+      closeBeforeAction: true
+    }
+  ];
+
+  return <ContextMenu menu={menu} contextMenu={contextMenu} onClose={onClose} />;
 }
